requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
torch==2.11.0
diffusers==0.37.1
transformers==4.49.0
pillow==12.2.0
safetensors==0.7.0
timm==1.0.26
einops==0.8.2
protobuf==6.32.1
